requirements.txt
babel==2.17.0
chardet==5.2.0
customtkinter==5.2.2
darkdetect==0.8.0
fpdf==1.7.2
mysql==0.0.3
mysql-connector-python==9.2.0
mysqlclient==2.2.7
packaging==24.2
pillow==11.1.0
reportlab==4.3.1
setuptools==75.8.0
tkcalendar==1.6.1
wheel==0.45.1